Beginning setup to connect to EventStore embedded.
Right now, the current state of the app is non-functional.  This is the start of using EventStore in Blazor.  Will need to coordinate with the EventStore team to see if they can re-tool the Grpc Client to be netstandard 2.0 or 2.1 to allow the use within Blazor, pending DNS resolution will not be an issue.

TCP and HTTP endpoints are being deprecated in favor of the GRPC protocol.